get the posix timers stuff fixed. Have ping cookies be stored in a u_int32_t
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@19431 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/struct.h
@@ -229,7 +229,7 @@ struct LocalUser
 	int actually_read;	/* how many we've actually read in this second */
 	int sent_parsed;	/* how many messages we've parsed in this second */
 	time_t last_knock;	/* time of last knock */
-	unsigned long random_ping;
+	u_int32_t random_ping;
 	struct AuthRequest	*auth_request;
 
 	/* target change stuff */
